requirements.txt
pandas==1.5.3
requests==2.28.2
google-cloud-storage==2.8.0
google-cloud-bigquery==3.10.0
